requirements.txt
langchain-core
pyyaml
pydantic
scikit-learn
# Core AI & Logic
langchain==0.3.*
langchain-community
langchain-google-genai==2.1.*
langgraph==0.2.*

# Database & Security
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
pydantic-settings==2.4.*

# UI & Data Viz
streamlit==1.38.*
plotly==5.24.*
pandas==2.2.*

# API (For later)
fastapi==0.112.*
uvicorn==0.30.*

#redis
redis==5.0.8
